Moving conditions service to services directory
